Isolate Rust XML helpers from common code
Keep shared Python generation code identical by moving Rust XML naming helpers into a Rust-specific module.


Copilot-Session: 36d8b1f7-31ac-481f-a23c-a042eef5c822
